requirements.txt
xgboost>=2.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
pyarrow>=14.0.0
pandera>=0.17.0
scipy>=1.11.0
pyyaml>=6.0
streamlit>=1.30.0
plotly>=5.18.0
pytest>=7.4.0
pytest-cov>=4.1.0